+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.1.1
   - FEATURE: AFI reported in the alarm JSON
+  - FIX: alarm/exec on clearing alarm
 
 # 0.1.0
   - wooo